refactored folder for FE
