+++ front-end/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
     collectCoverage: true,
-    collectCoverageFrom: ['src/components/*.{js,jsx}', 'src/services/*.{js,jsx}'],
+    collectCoverageFrom: ['src/components/**/*.{js,jsx}', 'src/services/*.{js,jsx}'],
     coverageDirectory: 'coverage',
     testEnvironment: 'jsdom',
     setUpFilesAfterEnv: ['src/jest.setup.js'],
