+++ front-end/src/components/analysis-page/AnalysisPage.test.js
@@ -2,7 +2,7 @@ import {render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
-import store from '../store.js';
+import store from '../../store.js';
 import AnalysisPage from './AnalysisPage.jsx';
 
 describe('AnalysisPage tests', () => {
