+++ front-end/src/components/analysis-page/IntentDiagram.jsx
