+++ front-end/src/components/analysis-page/IntentDiagram.test.js
