+++ front-end/src/components/general/Icon.test.js
