+++ front-end/src/components/main-page/MainPage.test.js
@@ -2,9 +2,9 @@ import {render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
-import store from '../store.js';
+import store from '../../store.js';
 import MainPage from './MainPage.jsx';
-import {setTranscriptUploadStatusTrue, setTranscriptUploadStatusFalse} from "../store/transcriptUploadSlice.js"
+import {setTranscriptUploadStatusTrue, setTranscriptUploadStatusFalse} from "../../store/transcriptUploadSlice.js";
 
 describe('MainPage tests', () => {
     const renderComponent = () => render(
